pyproject.toml
[tool.poetry]
name = "realmate-backend-challenge"
version = "0.1.0"
description = "Desafio técnico Realmate - Backend Pleno (API + Webhooks)"
authors = ["Samuel Lima <samuelamlima@hotmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
django-cors-headers = "^4.3"

[tool.poetry.scripts]
start = "manage:runserver"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
